reset password pushed test diff up to 100
related to #264
